+++ lib/scout_apm/agent.rb
@@ -5,8 +5,6 @@ module ScoutApm
   # The worker thread wakes up every +Agent#period+, merges in-memory metrics w/those saved to disk, 
   # saves tshe merged data to disk, and sends it to the Scout server.
   class Agent
-    # Headers passed up with all API requests.
-    HTTP_HEADERS = { "Agent-Hostname" => Socket.gethostname }
     # see self.instance
     @@instance = nil 
     
@@ -49,9 +47,12 @@ module ScoutApm
     def start(options = {})
       @options.merge!(options)
       init_logger
-      logger.info "Attempting to start Scout Agent [#{ScoutApm::VERSION}] on [#{Socket.gethostname}]"
-      if !config.settings['monitor'] and !@options[:force]
+      logger.info "Attempting to start Scout Agent [#{ScoutApm::VERSION}] on [#{environment.hostname}]"
+      if !config.value('monitor') and !@options[:force]
         logger.warn "Monitoring isn't enabled for the [#{environment.env}] environment."
+        return false
+      elsif !config.value('name')
+        logger.warn "An application name is required. Specify the :name value in scout_apm.yml. Not starting agent."
         return false
       elsif !environment.app_server
         logger.warn "Couldn't find a supported app server. Not starting agent."
@@ -61,7 +62,7 @@ module ScoutApm
         return false
       end
       @started = true
-      logger.info "Starting monitoring. Framework [#{environment.framework}] App Server [#{environment.app_server}]."
+      logger.info "Starting monitoring for [#{config.value('name')}]. Framework [#{environment.framework}] App Server [#{environment.app_server}]."
       start_instruments
       if !start_background_worker?
         logger.debug "Not starting worker thread"