+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 
 * Add support for specifying logging to STDOUT
 * Heroku environment detection
+* Environment vars with "SCOUT_" prefix override any settings specified in the config file.
+* If on Heroku uses the dyno name vs. the hostname as the hostname
 
 # 0.1.1